+++ aiohttp_sqlalchemy/__init__.py
@@ -43,7 +43,6 @@ from aiohttp_sqlalchemy.views import (
     SAModelView,
     SAModelViewMixin,
     SAPrimaryKeyMixin,
-    SAView,
 )
 
 __version__ = '0.26.0'
@@ -74,7 +73,6 @@ __all__ = [
     'SAMixin',
     'SAModelView',
     'SAPrimaryKeyMixin',
-    'SAView',
     'bind',
     'get_session',
     'get_session_factory',
@@ -84,7 +82,6 @@ __all__ = [
     'setup',
     # Synonyms
     'DEFAULT_KEY',
-    'SAView',
     'sa_bind',
     'sa_init_db',
     'sa_session',